Rebase, reorganize tests, separate tokenStorageBuilder
