+++ lib/TokenManager.js
@@ -1,34 +1,9 @@
 var util = require('./util');
 var AuthSdkError = require('./errors/AuthSdkError');
-var config = require('./config.json');
 var cookies = require('./cookies');
+var tokenStorageBuilder = require('./tokenStorageBuilder');
 
-function tokenStorageBuilder(storage) {
-  function getTokenStorage() {
-    var tokenStorageString = storage.getItem(config.TOKEN_STORAGE_NAME);
-    tokenStorageString = tokenStorageString || '{}';
-    try {
-      return JSON.parse(tokenStorageString);
-    } catch(e) {
-      throw new AuthSdkError('Unable to parse token storage string');
-    }
-  }
-
-  function setTokenStorage(tokenStorage) {
-    try {
-      var tokenStorageString = JSON.stringify(tokenStorage);
-      storage.setItem(config.TOKEN_STORAGE_NAME, tokenStorageString);
-    } catch(e) {
-      throw new AuthSdkError('Unable to set token storage string');
-    }
-  }
-
-  return {
-    getTokenStorage: getTokenStorage,
-    setTokenStorage: setTokenStorage
-  };
-}
-
+// Provides webStorage-like interface for cookies
 var cookieStorage = {
   getItem: cookies.getCookie,
   setItem: function(key, value) {
